feat: get price beforehand
